requirements.txt
easyocr==1.6.2
opencv-python==4.7.0.68
pyspellchecker==0.6.3